Update profile to populate default label for bclup subjects
