+++ src/bluecore/utils.js
@@ -100,6 +100,17 @@ export function applyBluecoreLookupRequest(url, options = {}) {
   return { url: resolvedUrl, options: requestOptions, cbd: isInstancePath }
 }
 
+export function updateBclupSourceInProfile(valueConstraint) {
+  for (const source of BCLUP_SOURCE) {
+    if (!valueConstraint.defaults.some((d) => d.defaultURI === source.uri)) {
+      valueConstraint.defaults.push({
+        defaultLiteral: source.label,
+        defaultURI: source.uri
+      });
+    }
+  }
+}
+
 export function generateBclupResultEntry(hit, length) {
   return {
     collections: [],
@@ -137,6 +148,16 @@ export function generateSuggestLabelPostfix(mode, gettySearchType = '') {
 
 export function isBclupMode(mode) {
   return mode == BCLUP_GETTY_MODE || mode == BCLUP_HOMOSAURUS_MODE
+}
+
+export function isBclupSource(uri) {
+  if (!uri || typeof uri !== 'string') return false
+  for (const source of BCLUP_SOURCE) {
+    if (uri.indexOf(source.prefix) > -1) {
+      return true
+    }
+  }
+  return false
 }
 
 export function buildEmptySubjectSearchResults() {
