+++ src/components/panels/edit/modals/helpers/ComplexSearchResultsDisplay.vue
@@ -93,6 +93,19 @@
                 @emitLoadContext="loadContext"
                 :searchMode="searchMode"
             />
+
+            <SearchResultOption
+                searchType="bclup"
+                :label="searchMode == 'BCLUP_HOMOSAURUS' ? 'Homosaurus' : 'Getty'"
+                index="ix"
+                :searchResults="searchResults"
+                :pickLookup="pickLookup"
+                @selectContext="selectContext"
+                @emitLoadContext="loadContext"
+                :searchMode="searchMode"
+                :gettySearchType="gettySearchType"
+                @lookupSearch="lookupSearch"
+            />
         </div>
     </div>
 </template>
@@ -141,6 +154,10 @@ export default {
         searchResults: Object,
         pickLookup: Object,
         searchMode: String,
+        gettySearchType: {
+            type: String,
+            default: 'aat',
+        },
     },
 
     watch: {},
@@ -166,6 +183,10 @@ export default {
 
         nafSearch: function(type){
             this.$emit('nafSearch', type)
+        },
+
+        lookupSearch: function(type){
+            this.$emit('lookupSearch', type)
         },
 
         loadContext:function(pickPosition){
