+++ src/components/panels/edit/modals/helpers/DetailsPanel.vue
@@ -37,7 +37,7 @@
                 <br>
             </template>
             <a style="color:#2c3e50" :href="rewriteURI(contextData.uri)" target="_blank"
-                v-if="contextData.literal != true">view on id.loc.gov</a>
+                v-if="contextData.literal != true">{{ getContextLinkLabel(contextData.uri) }}</a>
             <!-- Dates -->
             <template v-if="(Object.keys(contextData).includes('birthdates') && contextData['birthdates'].length > 0)
                 || (Object.keys(contextData).includes('deathdates') && contextData['deathdates'].length > 0)">
@@ -423,6 +423,15 @@ export default {
 
 
             return uri
+        },
+        getContextLinkLabel: function (uri) {
+            let rewrittenUri = this.rewriteURI(uri)
+
+            if (rewrittenUri && rewrittenUri.includes('id.loc.gov')) {
+                return 'view on id.loc.gov'
+            }
+
+            return 'view full description'
         },
     },
